管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -28,6 +28,8 @@ public class EmployeeAction extends ActionBase {
 
     public void index() throws ServletException, IOException {
 
+        if(checkAdmin()) {
+
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
 
@@ -45,19 +47,24 @@ public class EmployeeAction extends ActionBase {
         }
 
         forward(ForwardConst.FW_EMP_INDEX);
+        }
     }
 
     public void entryNew() throws ServletException, IOException {
+
+        if(checkAdmin()) {
 
         putRequestScope(AttributeConst.TOKEN, getTokenId());
         putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());
 
         forward(ForwardConst.FW_EMP_NEW);
+        }
+
     }
 
     public void create() throws ServletException, IOException {
 
-        if(checkToken()) {
+        if(checkAdmin() && checkToken()) {
             EmployeeView ev = new EmployeeView(
                     null,
                     getRequestParam(AttributeConst.EMP_CODE),
@@ -91,6 +98,8 @@ public class EmployeeAction extends ActionBase {
 
     public void show() throws ServletException, IOException {
 
+        if(checkAdmin()) {
+
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
         if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
@@ -102,9 +111,12 @@ public class EmployeeAction extends ActionBase {
         putRequestScope(AttributeConst.EMPLOYEE, ev);
 
         forward(ForwardConst.FW_EMP_SHOW);
+        }
     }
 
     public void edit() throws ServletException, IOException {
+
+        if(checkAdmin()) {
 
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -118,12 +130,13 @@ public class EmployeeAction extends ActionBase {
         putRequestScope(AttributeConst.EMPLOYEE, ev);
 
         forward(ForwardConst.FW_EMP_EDIT);
+        }
 
     }
 
     public void update() throws ServletException, IOException {
 
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) {
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
                     getRequestParam(AttributeConst.EMP_CODE),
@@ -156,13 +169,25 @@ public class EmployeeAction extends ActionBase {
 
     public void destroy() throws ServletException, IOException {
 
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) {
 
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
             putSessionScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
 
             redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+        }
+    }
+
+    private boolean checkAdmin() throws ServletException, IOException {
+
+        EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+        if(ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+        } else {
+            return true;
         }
     }
 }
